se implemento carousele en el home, se limpio mas codigo y se modificaron algunas cosas de algunos componentes para evitar las advertencias de react
