+++ .gitignore
@@ -23,3 +23,5 @@ package-lock.json
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+.env
