+++ src/components/Carousel.jsx
@@ -1,0 +1,29 @@
+import React from 'react'
+import banner1 from "../assets/Agregar un título.png"
+import banner2 from "../assets/TODO LO QUE NECESITAS (1).png"
+
+const Carousel = () => {
+  return (
+    <div id="carouselExample" className="carousel slide">
+  <div className="carousel-inner">
+    <div className="carousel-item active">
+      <img src={banner1} className="d-block w-100" alt="..." height="20%"/>
+    </div>
+    <div className="carousel-item">
+      <img src={banner2} className="d-block w-100" alt="..." />
+    </div>
+    
+  </div>
+  <button className="carousel-control-prev" type="button" data-bs-target="#carouselExample" data-bs-slide="prev">
+    <span className="carousel-control-prev-icon" aria-hidden="true"></span>
+    <span className="visually-hidden">Previous</span>
+  </button>
+  <button className="carousel-control-next" type="button" data-bs-target="#carouselExample" data-bs-slide="next">
+    <span className="carousel-control-next-icon" aria-hidden="true"></span>
+    <span className="visually-hidden">Next</span>
+  </button>
+</div>
+  )
+}
+
+export default Carousel
